Fix get_map_content and get_next_line related leaks
`get_next_line` was freeing line in inner scope when `ft_substr` failed. In normal situation sit should be fine to do so, but as we lose access to line pointer recieved, in outer scope we also lose information which is used to cleanly exit app.

When `ft_substr` doesn't fail, it's fine to free as we have new line without new linebreak. However, that's not the case when `ft_substr` fails and for that reason it was necessary to skip freeing `line`.