Fix missing exit (#73)
When player move from to tile that has coin inside and it's last coin, exit cannot be printed because of wrong order of printing. Fixed it by checking if current player position is exit (which is invisible) or not. If not, it prints floor, otherwise, `update_collected_coins` functions will handle it.